Attempt to Fix ArcadeDrive
WNM- I think I figured it out. After watching the video I moved this

"     	Robot.driveTrain.arcadeDrive(Robot.oi.getDriverXbox());
"
From initialize to execute. Which was like it was in the video.

Also let's test smart dashboard and see what we can get it to show that
is helpful.

Next up I am going to try and work on the high and low speed drive
modes.