Changes from UNH District event
-Commented out the Ultrasonic Sensor. We don't use it for anything and
we were making sure it would not take any CPU on the roboRIO. 

- Commented out the buttons for testing commands.

- We made some changes to autonomous. We would like to get the working
more. I think we need motor encoders to make it work correctly or some
way to tell the distance the robot has went. Using time is not accurate.
Need to figure out a way to do this without too much programming and
work.